chore: move container logs download to base class
